examples/camera-python-display/python/pyproject.toml
# Copyright (c) 2025 Jonathan Fontanez
# SPDX-License-Identifier: BUSL-1.1

[project]
name = "cyberpunk-processor"
version = "0.1.0"
description = "Cyberpunk video overlay processors using isolated subprocess architecture"
requires-python = ">=3.10"
dependencies = [
    "skia-python>=87.0",
    "PyOpenGL>=3.1.0",
    "mediapipe>=0.10.18",
    "numpy>=1.24.0",
    "moderngl>=5.8.0",
    "pygltflib>=1.16.0",
    "pyrr>=0.10.3",
    "pillow>=10.0.0",
    "msgpack>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.streamlib]
schemas = []
processors = [
    "avatar_character.yaml",
    "cyberpunk_lower_third.yaml",
    "cyberpunk_watermark.yaml",
    "cyberpunk_glitch.yaml",
]
